Added the IgnoreBrightness option for Analysis.
IgnoreBrightness: Determine the brightness value, when the brightness of the pixel exceeds this value, it is not detected, 1.0 represents white (Default 0.9)

Added a save feature in the main window, which can save the analysis result images and analysis result configuration *.yaml file.
Added support for opening analysis result configuration *.yaml file in the main window.

Improve the analysis mechanism.